+++ DiceJack/Helpers.h
@@ -1,4 +1,4 @@
 #pragma once
 
 int rollDie();
-int randomizeComputerBet(int min, double max);
+int randomizeComputerBet(double min, double max);
